refactor(mitm-addon): split sse usage tests by provider lifecycle
