--- crates/runner/mitm-addon/tests/test_model_provider_sse_usage.py
@@ -1,1341 +1,0 @@
-"""Tests for model-provider SSE usage reporting paths."""
-
-import gzip
-import json
-import uuid
-import zlib
-from collections.abc import Callable
-from datetime import UTC, datetime
-from pathlib import Path
-from unittest.mock import patch
-
-import brotli
-import pytest
-from mitmproxy import http
-from mitmproxy.flow import Error
-
-import anthropic_accounting
-import body_decoding
-import flow_metadata_keys as metadata_keys
-import mitm_addon
-import usage
-from tests.flow_helpers import response_stream
-from tests.jsonl_log_helpers import (
-    jsonl_exists_after_flush,
-    read_jsonl_entries_after_flush,
-)
-from tests.model_provider_flow_helpers import (
-    make_model_provider_sse_flow,
-    model_usage_source_entries,
-)
-from tests.pending_helpers import assert_current_pending, assert_pending
-from tests.usage_buffer_helpers import event as usage_event
-from tests.usage_helpers import (
-    CapturedWebhookRequest,
-    UsageWebhookServer,
-    compact_observation_quantities,
-)
-from tests.webhook_test_helpers import (
-    QueuedUsageExecutor,
-    install_runner_usage_flush_request,
-    request_runner_usage_flush,
-)
-from usage import openai_responses
-
-_TELEMETRY_PATH = "/api/webhooks/agent/telemetry"
-
-
-def _model_provider_sse_flow(
-    tmp_path: Path,
-    real_flow: Callable[..., http.HTTPFlow],
-    *,
-    host: str,
-    original_url: str,
-    firewall_name: str,
-    cli_agent_type: str | None = None,
-    model_usage_provider: str | None = "claude-sonnet-4-6",
-) -> http.HTTPFlow:
-    return make_model_provider_sse_flow(
-        real_flow,
-        tmp_path,
-        host=host,
-        original_url=original_url,
-        firewall_name=firewall_name,
-        cli_agent_type=cli_agent_type,
-        model_usage_provider=model_usage_provider,
-    )
-
-
-def _openai_responses_sse_flow(
-    tmp_path: Path,
-    real_flow: Callable[..., http.HTTPFlow],
-    *,
-    model_usage_provider: str = "gpt-5.5",
-) -> http.HTTPFlow:
-    return _model_provider_sse_flow(
-        tmp_path,
-        real_flow,
-        host="api.openai.com",
-        original_url="https://api.openai.com/v1/responses",
-        firewall_name="model-provider:openai-api-key",
-        cli_agent_type="codex",
-        model_usage_provider=model_usage_provider,
-    )
-
-
-def _anthropic_messages_sse_flow(
-    tmp_path: Path,
-    real_flow: Callable[..., http.HTTPFlow],
-) -> http.HTTPFlow:
-    flow = _model_provider_sse_flow(
-        tmp_path,
-        real_flow,
-        host="api.anthropic.com",
-        original_url="https://api.anthropic.com/v1/messages",
-        firewall_name="model-provider:anthropic-api-key",
-    )
-    flow.metadata[metadata_keys.VM_RUN_ID] = "00000000-0000-0000-0000-000000025133"
-    return flow
-
-
-def _compress_zlib_sse(body: bytes, encoding: str) -> bytes:
-    if encoding == "gzip":
-        return gzip.compress(body)
-    assert encoding == "deflate"
-    return zlib.compress(body)
-
-
-def _feed_incomplete_anthropic_sse_without_recoverable_usage(
-    flow: http.HTTPFlow,
-    *,
-    encoding: str = "gzip",
-) -> None:
-    assert flow.response is not None
-    flow.response.headers["content-encoding"] = encoding
-    plaintext = (
-        b"event: message_start\n"
-        b'data: {"type":"message_start","message":{"model":"claude-sonnet-4-6",'
-        b'"usage":{"input_tokens":0,"output_tokens":0}}}\n\n'
-        b"event: message_stop\n"
-        b'data: {"type":"message_stop"}\n\n'
-    )
-    mitm_addon.responseheaders(flow)
-    response_stream(flow)(_compress_zlib_sse(plaintext, encoding)[:-1])
-
-
-def _anthropic_accounting_requests(
-    webhook: UsageWebhookServer,
-) -> list[CapturedWebhookRequest]:
-    return [request for request in webhook.requests if request.path == _TELEMETRY_PATH]
-
-
-def _anthropic_accounting_operations(
-    webhook: UsageWebhookServer,
-) -> list[dict[str, object]]:
-    operations: list[dict[str, object]] = []
-    for request in _anthropic_accounting_requests(webhook):
-        request_operations = request.json_body().get("sandboxOperations")
-        assert isinstance(request_operations, list)
-        for operation in request_operations:
-            assert isinstance(operation, dict)
-            operations.append(operation)
-    return operations
-
-
-def _model_sse_parse_warnings(flow: http.HTTPFlow) -> list[dict]:
-    proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-    if not jsonl_exists_after_flush(proxy_log):
-        return []
-    return [
-        entry
-        for entry in read_jsonl_entries_after_flush(proxy_log)
-        if entry.get("message") == "Model provider SSE usage extraction failed"
-    ]
-
-
-def _assert_single_model_sse_parse_warning(
-    flow: http.HTTPFlow,
-    *,
-    usage_protocol: str,
-    event: str,
-    error: str | None = None,
-) -> None:
-    usage_warnings = _model_sse_parse_warnings(flow)
-    assert len(usage_warnings) == 1
-    warning = usage_warnings[0]
-    assert warning["level"] == "warn"
-    assert warning["type"] == "usage_event"
-    assert warning["usage_protocol"] == usage_protocol
-    assert warning["event"] == event
-    if error is None:
-        assert warning["error"]
-    else:
-        assert warning["error"] == error
-
-
-class TestModelProviderSseUsage:
-    """Tests for model-provider SSE usage reporting."""
-
-    @pytest.fixture(autouse=True)
-    def _sync_usage_delivery(self, sync_usage_executor, usage_webhook_api):
-        self._usage_webhook_api = usage_webhook_api
-
-    def _run_response(self, flow: http.HTTPFlow):
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.response(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def _run_error(self, flow: http.HTTPFlow):
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def test_full_pipeline_model_sse_finalizes_trailing_event(self, tmp_path, real_flow):
-        """response() must flush a trailing SSE usage event before reporting."""
-        flow = _openai_responses_sse_flow(
-            tmp_path,
-            real_flow,
-            model_usage_provider="gpt-5.6-sol",
-        )
-        mitm_addon.responseheaders(flow)
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-        response_stream(flow)(
-            b"event: response.completed\n"
-            b'data: {"response":{"model":"gpt-5.6-sol",'
-            b'"usage":{"input_tokens":50,"output_tokens":20,'
-            b'"input_tokens_details":{"cached_tokens":10,'
-            b'"cache_write_tokens":15}}}}'
-        )
-
-        webhook = self._run_response(flow)
-
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert len(events) == len(by_category)
-        assert by_category == {
-            "tokens.input": 25,
-            "tokens.output": 20,
-            "tokens.cache_read": 10,
-            "tokens.cache_creation": 15,
-        }
-
-    def test_full_pipeline_pi_responses_path_reports_usage(self, tmp_path, real_flow):
-        flow = _model_provider_sse_flow(
-            tmp_path,
-            real_flow,
-            host="api.deepseek.com",
-            original_url="https://api.deepseek.com/responses",
-            firewall_name="model-provider:deepseek",
-            cli_agent_type="pi",
-            model_usage_provider="deepseek-v4-flash",
-        )
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: response.completed\n"
-            b'data: {"type":"response.completed","response":{"model":"deepseek-chat",'
-            b'"usage":{"input_tokens":7131,"output_tokens":49,'
-            b'"input_tokens_details":{"cached_tokens":5504}}}}\n\n'
-        )
-
-        webhook = self._run_response(flow)
-
-        assert {event["category"]: event["quantity"] for event in webhook.usage_events()} == {
-            "tokens.input": 1627,
-            "tokens.output": 49,
-            "tokens.cache_read": 5504,
-        }
-
-    def test_full_pipeline_openai_sse_reports_long_context_items(self, tmp_path, real_flow):
-        flow = _openai_responses_sse_flow(
-            tmp_path,
-            real_flow,
-            model_usage_provider="gpt-5.6-sol",
-        )
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: response.completed\n"
-            b'data: {"response":{"model":"gpt-5.6-sol",'
-            b'"usage":{"input_tokens":272001,"output_tokens":20,'
-            b'"input_tokens_details":{"cached_tokens":70000,'
-            b'"cache_write_tokens":2001}}}}\n\n'
-        )
-
-        webhook = self._run_response(flow)
-
-        assert {event["category"]: event["quantity"] for event in webhook.usage_events()} == {
-            "tokens.input.long_context": 200_000,
-            "tokens.output.long_context": 20,
-            "tokens.cache_read.long_context": 70_000,
-            "tokens.cache_creation.long_context": 2_001,
-        }
-        assert compact_observation_quantities(webhook.model_usage_observation_events()) == {
-            "tokens.input": 200_000,
-            "tokens.output": 20,
-            "tokens.cache_read": 70_000,
-            "tokens.cache_creation": 2_001,
-        }
-
-    def test_full_pipeline_openai_sse_reports_usage_with_oversized_type(self, tmp_path, real_flow):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: response.completed\n"
-            b'data: {"type":"' + b"x" * 2048 + b'","response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":50,"output_tokens":20}}}\n\n'
-        )
-
-        webhook = self._run_response(flow)
-
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == {
-            "tokens.input": 50,
-            "tokens.output": 20,
-        }
-        assert _model_sse_parse_warnings(flow) == []
-
-    @pytest.mark.parametrize("capture_body", [False, True])
-    def test_brotli_sse_does_not_use_model_json_fallback(self, tmp_path, real_flow, capture_body):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "br"
-        if capture_body:
-            flow.metadata[metadata_keys.CAPTURE_BODY] = True
-        plaintext = (
-            b"event: response.completed\n"
-            b'data: {"response":{"id":"resp_sse_1","model":"gpt-5.5",'
-            b'"usage":{"input_tokens":50,"output_tokens":20}}}\n\n'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(brotli.compress(plaintext))
-        assert (metadata_keys.STREAM_BUFFER in flow.metadata) is capture_body
-        assert (metadata_keys.STREAM_BUFFER_STATE in flow.metadata) is capture_body
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-        entries = (
-            read_jsonl_entries_after_flush(proxy_log) if jsonl_exists_after_flush(proxy_log) else []
-        )
-        assert not any(
-            entry.get("message") == "Model provider JSON usage extraction failed"
-            for entry in entries
-        )
-
-    def test_full_pipeline_model_sse_reports_response_incomplete_usage(self, tmp_path, real_flow):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: response.incomplete\n"
-            b'data: {"response":{"id":"resp_incomplete","model":"gpt-5.5",'
-            b'"usage":{"input_tokens":8000,"output_tokens":1024,'
-            b'"input_tokens_details":{"cached_tokens":2000}}}}\n\n'
-        )
-
-        webhook = self._run_response(flow)
-
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == {
-            "tokens.input": 6000,
-            "tokens.output": 1024,
-            "tokens.cache_read": 2000,
-        }
-        assert {event["provider"] for event in events} == {"gpt-5.5"}
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    @pytest.mark.parametrize("hook_name", ["response", "error"])
-    def test_full_pipeline_incomplete_compressed_openai_sse_recovers_terminal_usage(
-        self, tmp_path, real_flow, encoding, hook_name
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = encoding
-        plaintext = (
-            b"event: response.completed\n"
-            b'data: {"response":{"id":"resp_sse_1","model":"gpt-5.5",'
-            b'"usage":{"input_tokens":100,"output_tokens":40,'
-            b'"input_tokens_details":{"cached_tokens":20,'
-            b'"cache_write_tokens":30}}}}\n\n'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(_compress_zlib_sse(plaintext, encoding)[:-1])
-        if hook_name == "error":
-            flow.error = Error("connection reset by peer")
-            webhook = self._run_error(flow)
-        else:
-            webhook = self._run_response(flow)
-
-        expected = {
-            "tokens.input": 50,
-            "tokens.output": 40,
-            "tokens.cache_read": 20,
-            "tokens.cache_creation": 30,
-        }
-        assert {
-            event["category"]: event["quantity"] for event in webhook.usage_events()
-        } == expected
-        assert compact_observation_quantities(webhook.model_usage_observation_events()) == expected
-        [source_entry] = model_usage_source_entries(flow)
-        assert source_entry["source_id"] == flow.id
-        assert source_entry["provider_response_id"] == "resp_sse_1"
-        assert source_entry["transport"] == "http"
-        assert source_entry["buffer_mode"] == "aggregate"
-        assert source_entry["usage"] == expected
-        assert all(event["buffer_accepted"] is True for event in source_entry["usage_events"])
-        assert all(
-            observation["buffer_accepted"] is True
-            for observation in source_entry["model_usage_observations"]
-        )
-        assert {
-            event["source_idempotency_key"] for event in source_entry["usage_events"]
-        }.isdisjoint({event["idempotencyKey"] for event in webhook.usage_events()})
-        assert {
-            observation["source_idempotency_key"]
-            for observation in source_entry["model_usage_observations"]
-        }.isdisjoint(
-            {
-                observation["idempotencyKey"]
-                for observation in webhook.model_usage_observation_events()
-            }
-        )
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="openai_responses_sse",
-            event="compressed_body",
-            error=body_decoding.INCOMPLETE_COMPRESSED_BODY,
-        )
-
-    @pytest.mark.parametrize(
-        ("event_prefix", "event_type"),
-        [
-            pytest.param(b"event: response.done\n", "response.done", id="done"),
-            pytest.param(
-                b"event: response.incomplete\n",
-                "response.incomplete",
-                id="incomplete",
-            ),
-            pytest.param(b"event: response.failed\n", "response.failed", id="failed"),
-            pytest.param(b"", "response.completed", id="eventless"),
-        ],
-    )
-    def test_full_pipeline_incomplete_compressed_openai_sse_recovers_terminal_vocabulary(
-        self, tmp_path, real_flow, event_prefix, event_type
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            event_prefix
-            + b'data: {"type":"'
-            + event_type.encode()
-            + b'","response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":11,"output_tokens":5}}}\n\n'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(gzip.compress(plaintext)[:-1])
-        webhook = self._run_response(flow)
-
-        assert {event["category"]: event["quantity"] for event in webhook.usage_events()} == {
-            "tokens.input": 11,
-            "tokens.output": 5,
-        }
-
-    @pytest.mark.parametrize("truncate_trailer", [False, True], ids=["complete", "incomplete"])
-    def test_full_pipeline_openai_recovery_excludes_unknown_event_usage(
-        self, tmp_path, real_flow, truncate_trailer
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            b"event: response.future_usage\n"
-            b'data: {"type":"response.future_usage","response":{"model":"gpt-5.5",'
-            b'"usage":{"output_tokens":99}}}\n\n'
-            b"event: response.completed\n"
-            b'data: {"type":"response.completed","response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":10}}}\n\n'
-        )
-        encoded = gzip.compress(plaintext)
-        if truncate_trailer:
-            encoded = encoded[:-1]
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(encoded)
-        webhook = self._run_response(flow)
-
-        expected = {"tokens.input": 10}
-        if not truncate_trailer:
-            expected["tokens.output"] = 99
-        assert {
-            event["category"]: event["quantity"] for event in webhook.usage_events()
-        } == expected
-
-    @pytest.mark.parametrize(
-        "type_fields",
-        [
-            pytest.param(b'"type":"response.failed",', id="sse-json-conflict"),
-            pytest.param(
-                b'"type":"response.completed","type":"response.failed",'
-                b'"type":"response.completed",',
-                id="middle-conflicts-with-matching-ends",
-            ),
-            pytest.param(
-                b'"padding":"'
-                + b"x" * (openai_responses._RESPONSES_EVENT_PREFILTER_MAX_BYTES + 1)
-                + b'","type":"response.failed","type":"response.completed",',
-                id="conflict-after-prefilter-bound",
-            ),
-            pytest.param(
-                b'"padding":"'
-                + b"x" * (openai_responses._RESPONSES_EVENT_PREFILTER_MAX_BYTES + 1)
-                + b'","type":42,',
-                id="invalid-type-after-prefilter-bound",
-            ),
-        ],
-    )
-    def test_full_pipeline_incomplete_compressed_openai_sse_rejects_conflicting_identity(
-        self, tmp_path, real_flow, type_fields
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            b"event: response.completed\n"
-            b"data: {" + type_fields + b'"response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":11,"output_tokens":5}}}\n\n'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(gzip.compress(plaintext)[:-1])
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-
-    def test_full_pipeline_incomplete_compressed_openai_sse_does_not_flush_partial_event(
-        self, tmp_path, real_flow
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            b"event: response.completed\n"
-            b'data: {"type":"response.completed","response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":11,"output_tokens":5}'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(gzip.compress(plaintext)[:-1])
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_full_pipeline_invalid_compressed_openai_sse_remains_fail_closed(
-        self, tmp_path, real_flow, encoding
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = encoding
-        plaintext = (
-            b"event: response.completed\n"
-            b'data: {"type":"response.completed","response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":11,"output_tokens":5}}}\n\n'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(_compress_zlib_sse(plaintext, encoding) + b"not-compressed")
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="openai_responses_sse",
-            event="compressed_body",
-            error=body_decoding.INVALID_COMPRESSED_BODY,
-        )
-
-    def test_full_pipeline_decoded_limit_openai_sse_remains_fail_closed(self, tmp_path, real_flow):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            b"event: response.completed\n"
-            b'data: {"type":"response.completed","response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":11,"output_tokens":5}}}\n\n'
-            b"event: response.output_text.delta\n"
-            b"data: " + b"x" * (5 * 1024 * 1024 + 1)
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(gzip.compress(plaintext))
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="openai_responses_sse",
-            event="compressed_body",
-            error=body_decoding.DECODED_BODY_LIMIT_EXCEEDED,
-        )
-
-    def test_full_pipeline_response_then_error_emits_recovered_openai_usage_once(
-        self, tmp_path, real_flow
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            b"event: response.completed\n"
-            b'data: {"response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":11,"output_tokens":5}}}\n\n'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(gzip.compress(plaintext)[:-1])
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.response(flow)
-            flow.error = Error("connection reset after response")
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-
-        assert [(event["category"], event["quantity"]) for event in webhook.usage_events()] == [
-            ("tokens.input", 11),
-            ("tokens.output", 5),
-        ]
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    @pytest.mark.parametrize("hook_name", ["response", "error"])
-    @pytest.mark.parametrize(
-        ("include_message_stop", "expected_status"),
-        [
-            (False, "recovered_partial"),
-            (True, "recovered_terminal"),
-        ],
-        ids=["partial", "terminal"],
-    )
-    def test_full_pipeline_incomplete_compressed_anthropic_sse_recovers_complete_usage_events(
-        self,
-        tmp_path,
-        real_flow,
-        encoding,
-        hook_name,
-        include_message_stop,
-        expected_status,
-    ):
-        flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = encoding
-        plaintext = (
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":{"id":"msg_private",'
-            b'"model":"claude-sonnet-4-6","usage":{"input_tokens":101,'
-            b'"cache_read_input_tokens":202,"cache_creation_input_tokens":303,'
-            b'"output_tokens":1}}}\n\n'
-            b"event: content_block_delta\n"
-            b'data: {"type":"content_block_delta","delta":{"text":"sensitive-body"}}\n\n'
-            b"event: message_delta\n"
-            b'data: {"type":"message_delta","delta":{"stop_reason":"end_turn"},'
-            b'"usage":{"output_tokens":404}}\n\n'
-            + (
-                b'event: message_stop\ndata: {"type":"message_stop"}\n\n'
-                if include_message_stop
-                else b""
-            )
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(_compress_zlib_sse(plaintext, encoding)[:-1])
-
-        if hook_name == "error":
-            flow.error = Error("connection reset by peer")
-            webhook = self._run_error(flow)
-        else:
-            assert hook_name == "response"
-            webhook = self._run_response(flow)
-
-        expected_quantities = {
-            "tokens.input": 101,
-            "tokens.output": 404,
-            "tokens.cache_read": 202,
-            "tokens.cache_creation": 303,
-        }
-        assert {
-            event["category"]: event["quantity"] for event in webhook.usage_events()
-        } == expected_quantities
-        assert (
-            compact_observation_quantities(webhook.model_usage_observation_events())
-            == expected_quantities
-        )
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="compressed_body",
-            error=body_decoding.INCOMPLETE_COMPRESSED_BODY,
-        )
-
-        [request] = _anthropic_accounting_requests(webhook)
-        assert request.header("authorization") == "Bearer tok-xyz"
-        payload = request.json_body()
-        assert payload["runId"] == "00000000-0000-0000-0000-000000025133"
-        [operation] = _anthropic_accounting_operations(webhook)
-        assert set(operation) == {
-            "ts",
-            "action_type",
-            "duration_ms",
-            "success",
-            "error",
-        }
-        assert operation["action_type"] == f"anthropic_sse_incomplete_{expected_status}"
-        assert operation["duration_ms"] == 0
-        assert operation["success"] is False
-        assert operation["error"] == body_decoding.INCOMPLETE_COMPRESSED_BODY
-        assert datetime.fromisoformat(str(operation["ts"])).tzinfo == UTC
-        serialized_request = request.body
-        for excluded in [
-            b"msg_private",
-            b"claude-sonnet-4-6",
-            b"sensitive-body",
-            b"tokens.",
-        ]:
-            assert excluded not in serialized_request
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_full_pipeline_incomplete_anthropic_sse_emits_no_recoverable_usage_status(
-        self, tmp_path, real_flow, encoding
-    ):
-        flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        _feed_incomplete_anthropic_sse_without_recoverable_usage(
-            flow,
-            encoding=encoding,
-        )
-        webhook = self._run_response(flow)
-
-        assert webhook.usage_events() == []
-        assert webhook.model_usage_observation_events() == []
-        [operation] = _anthropic_accounting_operations(webhook)
-        assert operation["action_type"] == "anthropic_sse_incomplete_no_recoverable_usage"
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="compressed_body",
-            error=body_decoding.INCOMPLETE_COMPRESSED_BODY,
-        )
-
-    def test_saturated_incomplete_anthropic_accounting_retries_through_runner_flush(
-        self,
-        tmp_path: Path,
-        real_flow,
-        mitm_ctx,
-        usage_webhook_server: UsageWebhookServer,
-    ) -> None:
-        flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        executor = QueuedUsageExecutor()
-        pending_path = install_runner_usage_flush_request(tmp_path)
-
-        with (
-            mitm_ctx(api_url=usage_webhook_server.api_url),
-            patch.object(usage.webhook, "usage_executor", executor),
-        ):
-            for index in range(usage.webhook.MAX_PENDING_WEBHOOK_PAYLOADS):
-                assert usage.webhook.enqueue_webhook_delivery(
-                    usage_webhook_server.url("/filler"),
-                    "tok-xyz",
-                    {"runId": f"filler-{index}", "events": []},
-                    str(tmp_path / "filler.jsonl"),
-                    "usage_event",
-                )
-
-            _feed_incomplete_anthropic_sse_without_recoverable_usage(flow)
-            mitm_addon.response(flow)
-            retained_at = datetime.now(UTC)
-
-            assert _anthropic_accounting_requests(usage_webhook_server) == []
-            assert_current_pending(
-                pending_path,
-                flows=0,
-                buffered=1,
-                reports=usage.webhook.MAX_PENDING_WEBHOOK_PAYLOADS,
-            )
-
-            assert (
-                usage.buffer_usage_events(
-                    usage_webhook_server.url("/usage-priority"),
-                    "tok-xyz",
-                    "usage-priority",
-                    [usage_event(source_key="usage-priority")],
-                    str(tmp_path / "proxy.jsonl"),
-                )
-                == 1
-            )
-            request_runner_usage_flush()
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=2,
-                reports=usage.webhook.MAX_PENDING_WEBHOOK_PAYLOADS,
-                flush_request_id="request-1",
-            )
-
-            executor.run_next()
-            request_runner_usage_flush()
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=2,
-                reports=usage.webhook.MAX_PENDING_WEBHOOK_PAYLOADS,
-                flush_request_id="request-1",
-            )
-
-            executor.run_last()
-            request_runner_usage_flush()
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=0,
-                reports=usage.webhook.MAX_PENDING_WEBHOOK_PAYLOADS,
-                flush_request_id="request-1",
-            )
-
-            executor.run_all()
-            request_runner_usage_flush()
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=0,
-                reports=0,
-                flush_request_id="request-1",
-            )
-            request_runner_usage_flush()
-
-        [request] = _anthropic_accounting_requests(usage_webhook_server)
-        assert request.json_body()["runId"] == "00000000-0000-0000-0000-000000025133"
-        [operation] = _anthropic_accounting_operations(usage_webhook_server)
-        assert operation["action_type"] == "anthropic_sse_incomplete_no_recoverable_usage"
-        assert datetime.fromisoformat(str(operation["ts"])) <= retained_at
-        assert [
-            captured.path
-            for captured in usage_webhook_server.requests
-            if captured.path != "/filler"
-        ] == ["/usage-priority", _TELEMETRY_PATH]
-        assert usage.webhook.pending_delivery_payload_count_for_tests() == 0
-
-    def test_incomplete_anthropic_accounting_retention_overflow_is_action_specific(
-        self,
-        tmp_path: Path,
-        real_flow,
-        mitm_ctx,
-        usage_webhook_server: UsageWebhookServer,
-    ) -> None:
-        first_flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        first_flow.metadata[metadata_keys.VM_RUN_ID] = "run-retained"
-        overflow_flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        overflow_flow.metadata[metadata_keys.VM_RUN_ID] = "run-overflow"
-        executor = QueuedUsageExecutor()
-        pending_path = tmp_path / "usage-pending"
-        usage.set_pending_path(str(pending_path))
-
-        with (
-            mitm_ctx(api_url=usage_webhook_server.api_url),
-            patch.object(usage.webhook, "usage_executor", executor),
-            patch.object(anthropic_accounting, "MAX_RETAINED_REPORTS", 1),
-        ):
-            for index in range(usage.webhook.MAX_PENDING_WEBHOOK_PAYLOADS):
-                assert usage.webhook.enqueue_webhook_delivery(
-                    usage_webhook_server.url("/filler"),
-                    "tok-xyz",
-                    {"runId": f"filler-{index}", "events": []},
-                    str(tmp_path / "filler.jsonl"),
-                    "usage_event",
-                )
-
-            _feed_incomplete_anthropic_sse_without_recoverable_usage(first_flow)
-            mitm_addon.response(first_flow)
-            _feed_incomplete_anthropic_sse_without_recoverable_usage(overflow_flow)
-            mitm_addon.response(overflow_flow)
-
-            assert_current_pending(
-                pending_path,
-                flows=0,
-                buffered=1,
-                reports=usage.webhook.MAX_PENDING_WEBHOOK_PAYLOADS,
-            )
-            overflow_entries = [
-                entry
-                for entry in read_jsonl_entries_after_flush(
-                    Path(overflow_flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-                )
-                if entry.get("reason") == "anthropic_accounting_retention_saturated"
-            ]
-            [overflow_entry] = overflow_entries
-            assert overflow_entry["type"] == "usage_underbilling"
-            assert overflow_entry["underbilling_class"] == "risk"
-            assert overflow_entry["component"] == "mitm_addon"
-            assert overflow_entry["action_type"] == (
-                "anthropic_sse_incomplete_no_recoverable_usage"
-            )
-            assert overflow_entry["run_id"] == "run-overflow"
-            assert overflow_entry["retained_report_capacity"] == 1
-            assert "tok-xyz" not in json.dumps(overflow_entry)
-
-            anthropic_accounting.reset_for_tests()
-            assert_current_pending(
-                pending_path,
-                flows=0,
-                buffered=0,
-                reports=usage.webhook.MAX_PENDING_WEBHOOK_PAYLOADS,
-            )
-            executor.run_all()
-
-        assert_current_pending(
-            pending_path,
-            flows=0,
-            buffered=0,
-            reports=0,
-        )
-        assert _anthropic_accounting_requests(usage_webhook_server) == []
-        assert usage.webhook.pending_delivery_payload_count_for_tests() == 0
-
-    def test_full_pipeline_incomplete_compressed_anthropic_sse_does_not_flush_fragment(
-        self, tmp_path, real_flow
-    ):
-        flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":{"model":"claude-sonnet-4-6",'
-            b'"usage":{"input_tokens":50,"output_tokens":0}}}\n\n'
-            b"event: message_delta\n"
-            b'data: {"type":"message_delta","usage":{"output_tokens":987654}}'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(gzip.compress(plaintext)[:-1])
-        webhook = self._run_response(flow)
-
-        assert {event["category"]: event["quantity"] for event in webhook.usage_events()} == {
-            "tokens.input": 50
-        }
-        assert compact_observation_quantities(webhook.model_usage_observation_events()) == {
-            "tokens.input": 50
-        }
-        [operation] = _anthropic_accounting_operations(webhook)
-        assert operation["action_type"] == "anthropic_sse_incomplete_recovered_partial"
-        assert b"987654" not in _anthropic_accounting_requests(webhook)[0].body
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_full_pipeline_invalid_compressed_anthropic_sse_remains_fail_closed(
-        self, tmp_path, real_flow, encoding
-    ):
-        flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = encoding
-        plaintext = (
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":{"model":"claude-sonnet-4-6",'
-            b'"usage":{"input_tokens":50}}}\n\n'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(_compress_zlib_sse(plaintext, encoding) + b"not-compressed")
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert _anthropic_accounting_operations(webhook) == []
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="compressed_body",
-            error=body_decoding.INVALID_COMPRESSED_BODY,
-        )
-
-    def test_full_pipeline_decoded_limit_anthropic_sse_remains_fail_closed(
-        self, tmp_path, real_flow
-    ):
-        flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":{"model":"claude-sonnet-4-6",'
-            b'"usage":{"input_tokens":50}}}\n\n'
-            b"event: content_block_delta\n"
-            b"data: " + b"x" * (5 * 1024 * 1024 + 1)
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(gzip.compress(plaintext))
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert _anthropic_accounting_operations(webhook) == []
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="compressed_body",
-            error=body_decoding.DECODED_BODY_LIMIT_EXCEEDED,
-        )
-
-    def test_full_pipeline_response_then_error_emits_recovered_usage_and_telemetry_once(
-        self, tmp_path, real_flow, mitm_ctx
-    ):
-        flow = _anthropic_messages_sse_flow(tmp_path, real_flow)
-        assert flow.response is not None
-        flow.response.headers["content-encoding"] = "gzip"
-        plaintext = (
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":{"model":"claude-sonnet-4-6",'
-            b'"usage":{"input_tokens":50}}}\n\n'
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(gzip.compress(plaintext)[:-1])
-        webhook = UsageWebhookServer()
-        with webhook.run(), mitm_ctx(api_url=webhook.api_url):
-            mitm_addon.response(flow)
-            flow.error = Error("connection reset after response")
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-
-        assert [(event["category"], event["quantity"]) for event in webhook.usage_events()] == [
-            ("tokens.input", 50)
-        ]
-        assert compact_observation_quantities(webhook.model_usage_observation_events()) == {
-            "tokens.input": 50
-        }
-        assert [
-            operation["action_type"] for operation in _anthropic_accounting_operations(webhook)
-        ] == ["anthropic_sse_incomplete_recovered_partial"]
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="compressed_body",
-            error=body_decoding.INCOMPLETE_COMPRESSED_BODY,
-        )
-
-    def test_full_pipeline_anthropic_sse_logs_truncated_message_start(self, tmp_path, real_flow):
-        flow = _model_provider_sse_flow(
-            tmp_path,
-            real_flow,
-            host="api.anthropic.com",
-            original_url="https://api.anthropic.com/v1/messages",
-            firewall_name="model-provider:anthropic-api-key",
-        )
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b'event: message_start\ndata: {"type":"message_start","message":{"id":"msg_1","mod'
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="message_start",
-        )
-
-    def test_full_pipeline_anthropic_sse_error_logs_truncated_message_start(
-        self, tmp_path, real_flow
-    ):
-        flow = _model_provider_sse_flow(
-            tmp_path,
-            real_flow,
-            host="api.anthropic.com",
-            original_url="https://api.anthropic.com/v1/messages",
-            firewall_name="model-provider:anthropic-api-key",
-        )
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b'event: message_start\ndata: {"type":"message_start","message":{"id":"msg_1","mod'
-        )
-        flow.error = Error("connection reset by peer")
-
-        webhook = self._run_error(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="message_start",
-        )
-
-    def test_full_pipeline_anthropic_sse_logs_malformed_message_start(self, tmp_path, real_flow):
-        flow = _model_provider_sse_flow(
-            tmp_path,
-            real_flow,
-            host="api.anthropic.com",
-            original_url="https://api.anthropic.com/v1/messages",
-            firewall_name="model-provider:anthropic-api-key",
-        )
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(b"event: message_start\ndata: {invalid json}\n\n")
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="message_start",
-        )
-
-    def test_full_pipeline_anthropic_sse_logs_truncated_message_delta_after_start(
-        self, tmp_path, real_flow
-    ):
-        flow = _model_provider_sse_flow(
-            tmp_path,
-            real_flow,
-            host="api.anthropic.com",
-            original_url="https://api.anthropic.com/v1/messages",
-            firewall_name="model-provider:anthropic-api-key",
-        )
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":{"id":"msg_1",'
-            b'"model":"claude-sonnet-4-6","usage":{"input_tokens":50}}}\n\n'
-            b"event: message_delta\n"
-            b'data: {"type":"message_delta","usage":{"output_tokens":'
-        )
-
-        webhook = self._run_response(flow)
-
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == {"tokens.input": 50}
-        assert {event["provider"] for event in events} == {"claude-sonnet-4-6"}
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="message_delta",
-        )
-
-    def test_full_pipeline_openai_sse_logs_truncated_terminal_event(self, tmp_path, real_flow):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: response.completed\n"
-            b'data: {"type":"response.completed","response":{"id":"resp_1","model":"gpt'
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="openai_responses_sse",
-            event="response.completed",
-        )
-
-    def test_full_pipeline_openai_sse_logs_truncated_late_event_name(self, tmp_path, real_flow):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b'data: {"type":"response.completed","response":{"id":"resp_1","model":"gpt\n'
-            b"event: response.completed\n\n"
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="openai_responses_sse",
-            event="response.completed",
-        )
-
-    def test_full_pipeline_eventless_incomplete_anthropic_usage_sse_warns(
-        self, tmp_path, real_flow
-    ):
-        flow = _model_provider_sse_flow(
-            tmp_path,
-            real_flow,
-            host="api.anthropic.com",
-            original_url="https://api.anthropic.com/v1/messages",
-            firewall_name="model-provider:anthropic-api-key",
-        )
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b'data: {"type":"message_start","message":{"id":"msg_1","model":"claude'
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="anthropic_messages_sse",
-            event="message_start",
-        )
-
-    def test_full_pipeline_anthropic_non_usage_incomplete_sse_does_not_warn(
-        self, tmp_path, real_flow
-    ):
-        flow = _model_provider_sse_flow(
-            tmp_path,
-            real_flow,
-            host="api.anthropic.com",
-            original_url="https://api.anthropic.com/v1/messages",
-            firewall_name="model-provider:anthropic-api-key",
-        )
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: content_block_delta\n"
-            b'data: {"type":"content_block_delta","delta":{"text":"hello'
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert _model_sse_parse_warnings(flow) == []
-
-    @pytest.mark.parametrize("hook_name", ["response", "error"])
-    def test_full_pipeline_openai_eventless_incomplete_terminal_sse_warns(
-        self, tmp_path, real_flow, hook_name
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b'data: {"type":"response.completed","response":{"id":"resp_1","model":"gpt'
-        )
-
-        if hook_name == "error":
-            flow.error = Error("connection reset by peer")
-            webhook = self._run_error(flow)
-        else:
-            webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        _assert_single_model_sse_parse_warning(
-            flow,
-            usage_protocol="openai_responses_sse",
-            event="response.completed",
-        )
-
-    def test_full_pipeline_openai_non_terminal_incomplete_sse_does_not_warn(
-        self, tmp_path, real_flow
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: response.in_progress\n"
-            b'data: {"type":"response.in_progress","response":{"id":"resp_1","model":"gpt'
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert _model_sse_parse_warnings(flow) == []
-
-    def test_full_pipeline_model_sse_zero_event_preserves_billed_usage_and_id(
-        self, tmp_path, real_flow
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b"event: response.completed\n"
-            b'data: {"response":{"id":"resp_sse_1","model":"gpt-5.5",'
-            b'"usage":{"input_tokens":100,"output_tokens":40}}}\n\n'
-            b"event: response.failed\n"
-            b'data: {"response":{"id":"resp_sse_empty","model":"gpt-5.6-luna",'
-            b'"usage":{"input_tokens":0,"output_tokens":0}}}\n\n'
-        )
-
-        webhook = self._run_response(flow)
-
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        idempotency_by_category = {event["category"]: event["idempotencyKey"] for event in events}
-        assert by_category == {
-            "tokens.input": 100,
-            "tokens.output": 40,
-        }
-        assert set(idempotency_by_category) == {"tokens.input", "tokens.output"}
-        for key in idempotency_by_category.values():
-            uuid.UUID(key)
-        assert {event["provider"] for event in events} == {"gpt-5.5"}
-
-    @pytest.mark.parametrize(
-        "later_cache_write_tokens",
-        [0, None],
-        ids=["zero", "omitted"],
-    )
-    def test_full_pipeline_model_sse_raw_snapshots_preserve_input_partition(
-        self,
-        tmp_path,
-        real_flow,
-        later_cache_write_tokens,
-    ):
-        flow = _openai_responses_sse_flow(tmp_path, real_flow)
-        mitm_addon.responseheaders(flow)
-        later_input_details = {"cached_tokens": 20}
-        if later_cache_write_tokens is not None:
-            later_input_details["cache_write_tokens"] = later_cache_write_tokens
-        first_snapshot = {
-            "type": "response.completed",
-            "response": {
-                "id": "resp_sse_partition",
-                "model": "gpt-5.5",
-                "usage": {
-                    "input_tokens": 100,
-                    "output_tokens": 0,
-                    "input_tokens_details": {
-                        "cached_tokens": 20,
-                        "cache_write_tokens": 30,
-                    },
-                },
-            },
-        }
-        later_snapshot = {
-            "type": "response.done",
-            "response": {
-                "id": "resp_sse_partition",
-                "model": "gpt-5.5",
-                "usage": {
-                    "input_tokens": 100,
-                    "output_tokens": 40,
-                    "input_tokens_details": later_input_details,
-                },
-            },
-        }
-        response_stream(flow)(
-            b"event: response.completed\n"
-            + b"data: "
-            + json.dumps(first_snapshot).encode()
-            + b"\n\n"
-            + b"event: response.done\n"
-            + b"data: "
-            + json.dumps(later_snapshot).encode()
-            + b"\n\n"
-        )
-
-        webhook = self._run_response(flow)
-
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert len(events) == len(by_category) == 4
-        assert by_category == {
-            "tokens.input": 50,
-            "tokens.output": 40,
-            "tokens.cache_read": 20,
-            "tokens.cache_creation": 30,
-        }
-        assert (
-            by_category["tokens.input"]
-            + by_category["tokens.cache_read"]
-            + by_category["tokens.cache_creation"]
-            == 100
-        )
-        observation_events = webhook.model_usage_observation_events()
-        observations_by_category = compact_observation_quantities(observation_events)
-        assert len(observation_events) == 1
-        assert observations_by_category == by_category
